# backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Interview AI Platform API
  version: 1.0.0
  description: API documentation for the Interview AI MVP backend.
servers:
  - url: http://localhost:5001/api
paths:
...
  /auth/register:
    post:
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
  /auth/login:
    post:
      summary: Login and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
...
  /auth/reset-password:
    post:
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password successfully reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseResponse"

  # ---------------------
  # USER
  # ---------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    BaseResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          nullable: true
        error:
          nullable: true

    # AUTH SCHEMAS
    RegisterRequest:
      type: object
      required: [email, password, full_name]
      properties:
        email:
          type: string
        password:
          type: string
        full_name:
          type: string
    RegisterResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            data:
              type: object
              properties:
                user_id:
                  type: string
                email:
                  type: string
                full_name:
                  type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    LoginResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            data:
              type: object
              properties:
                access_token:
                  type: string
                expires_in:
                  type: integer
...
    ResetPasswordRequest:
      type: object
      required: [reset_token, new_password]
      properties:
        reset_token:
          type: string
        new_password:
          type: string

    # USER SCHEMAS
...